+++ plugins/CustomVariables/CustomVariables.php
@@ -20,7 +20,6 @@ class CustomVariables extends \Piwik\Plugin
     public function registerEvents()
     {
         return array(
-            'Live.getAllVisitorDetails'       => 'extendVisitorDetails',
             'AssetManager.getJavaScriptFiles' => 'getJsFiles',
             'Translate.getClientSideTranslationKeys' => 'getClientSideTranslationKeys',
             'AssetManager.getStylesheetFiles'  => 'getStylesheetFiles',
@@ -35,24 +34,6 @@ class CustomVariables extends \Piwik\Plugin
     public function uninstall()
     {
         Model::uninstall();
-    }
-
-    public function extendVisitorDetails(&$visitor, $details)
-    {
-        $customVariables = array();
-
-        $maxCustomVariables = self::getNumUsableCustomVariables();
-
-        for ($i = 1; $i <= $maxCustomVariables; $i++) {
-            if (!empty($details['custom_var_k' . $i])) {
-                $customVariables[$i] = array(
-                    'customVariableName' .  $i => $details['custom_var_k' . $i],
-                    'customVariableValue' . $i => $details['custom_var_v' . $i],
-                );
-            }
-        }
-
-        $visitor['customVariables'] = $customVariables;
     }
 
     /**