+++ plugins/Live/Visitor.php
@@ -8,15 +8,15 @@
  */
 namespace Piwik\Plugins\Live;
 
+use Piwik\Cache;
+use Piwik\CacheId;
 use Piwik\Common;
 use Piwik\DataTable\Filter\ColumnDelete;
 use Piwik\Date;
-use Piwik\Db;
 use Piwik\Metrics\Formatter;
-use Piwik\Network\IPUtils;
+use Piwik\Plugin;
 use Piwik\Piwik;
 use Piwik\Plugins\CustomVariables\CustomVariables;
-use Piwik\Plugins\UserCountry\LocationProvider\GeoIp;
 use Piwik\Plugins\Actions\Actions\ActionSiteSearch;
 use Piwik\Tracker;
 use Piwik\Tracker\Action;
@@ -35,24 +35,14 @@ class Visitor implements VisitorInterface
 
     function getAllVisitorDetails()
     {
-        $visitor = array(
-            'idSite'                      => $this->getIdSite(),
-            'idVisit'                     => $this->getIdVisit(),
-            'visitIp'                     => $this->getIp(),
-            'visitorId'                   => $this->getVisitorId(),
+        $visitor = array();
 
-            // => false are placeholders to be filled in API later
-            'actionDetails'               => false,
-            'goalConversions'             => false,
-            'siteCurrency'                => false,
-            'siteCurrencySymbol'          => false,
+        $instances = $this->getAllVisitorDetailsInstances();
 
-            // all time entries
-            'serverDate'                  => $this->getServerDate(),
-            'visitServerHour'             => $this->getVisitServerHour(),
-            'lastActionTimestamp'         => $this->getTimestampLastAction(),
-            'lastActionDateTime'          => $this->getDateTimeLastAction(),
-        );
+        foreach ($instances as $instance) {
+            $instance->setDetails($this->details);
+            $instance->extendVisitorDetails($visitor);
+        }
 
         /**
          * This event can be used to add any details to a visitor. The visitor's details are for instance used in
@@ -70,10 +60,55 @@ class Visitor implements VisitorInterface
          *
          * @param array &visitor You can add or remove fields to the visitor array and it will reflected in the API output
          * @param array $details The details array contains all visit dimensions (columns of log_visit table)
+         *
+         * @deprecated  will be removed in Piwik 4
          */
         Piwik::postEvent('Live.getAllVisitorDetails', array(&$visitor, $this->details));
 
         return $visitor;
+    }
+
+    /**
+     * Returns all available activities
+     *
+     * @return VisitorDetailsAbstract[]
+     * @throws \Exception
+     */
+    protected function getAllVisitorDetailsInstances()
+    {
+        $cacheId = CacheId::pluginAware('VisitorDetails');
+        $cache   = Cache::getTransientCache();
+
+        if (!$cache->contains($cacheId)) {
+            $instances = [
+                new VisitorDetails() // needs to be first
+            ];
+
+            foreach ($this->getAllVisitorDetailsClasses() as $className) {
+                $instance = new $className();
+
+                if ($instance instanceof VisitorDetails) {
+                    continue;
+                }
+
+                $instances[] = $instance;
+            }
+
+            $cache->save($cacheId, $instances);
+        }
+
+        return $cache->fetch($cacheId);
+    }
+
+    /**
+     * Returns class names of all VisitorDetails classes.
+     *
+     * @return string[]
+     * @api
+     */
+    protected function getAllVisitorDetailsClasses()
+    {
+        return Plugin\Manager::getInstance()->findComponents('VisitorDetails', 'Piwik\Plugins\Live\VisitorDetailsAbstract');
     }
 
     function getVisitorId()
@@ -82,44 +117,6 @@ class Visitor implements VisitorInterface
             return bin2hex($this->details['idvisitor']);
         }
         return false;
-    }
-
-    function getVisitServerHour()
-    {
-        return date('G', strtotime($this->details['visit_last_action_time']));
-    }
-
-    function getServerDate()
-    {
-        return date('Y-m-d', strtotime($this->details['visit_last_action_time']));
-    }
-
-    function getIp()
-    {
-        if (isset($this->details['location_ip'])) {
-            return IPUtils::binaryToStringIP($this->details['location_ip']);
-        }
-        return null;
-    }
-
-    function getIdVisit()
-    {
-        return $this->details['idvisit'];
-    }
-
-    function getIdSite()
-    {
-        return $this->details['idsite'];
-    }
-
-    function getTimestampLastAction()
-    {
-        return strtotime($this->details['visit_last_action_time']);
-    }
-
-    function getDateTimeLastAction()
-    {
-        return date('Y-m-d H:i:s', strtotime($this->details['visit_last_action_time']));
     }
 
     /**