+++ plugins/Provider/Provider.php
@@ -17,16 +17,6 @@ use Piwik\Piwik;
 
 class Provider extends \Piwik\Plugin
 {
-    /**
-     * @see Piwik\Plugin::registerEvents
-     */
-    public function registerEvents()
-    {
-        return array(
-            'Live.getAllVisitorDetails' => 'extendVisitorDetails'
-        );
-    }
-
     public function install()
     {
         // add column hostname / hostname ext in the visit table
@@ -40,15 +30,6 @@ class Provider extends \Piwik\Plugin
                 throw $e;
             }
         }
-    }
-
-    public function extendVisitorDetails(&$visitor, $details)
-    {
-        $instance = new Visitor($details);
-
-        $visitor['provider']     = $instance->getProvider();
-        $visitor['providerName'] = $instance->getProviderName();
-        $visitor['providerUrl']  = $instance->getProviderUrl();
     }
 
     public function uninstall()