+++ plugins/Resolution/VisitorDetails.php
@@ -2,22 +2,22 @@
 /**
  * Piwik - free/libre analytics platform
  *
- * @link http://piwik.org
+ * @link    http://piwik.org
  * @license http://www.gnu.org/licenses/gpl-3.0.html GPL v3 or later
  *
  */
 namespace Piwik\Plugins\Resolution;
 
-class Visitor
-{
-    private $details = array();
+use Piwik\Plugins\Live\VisitorDetailsAbstract;
 
-    public function __construct($details)
+class VisitorDetails extends VisitorDetailsAbstract
+{
+    public function extendVisitorDetails(&$visitor)
     {
-        $this->details = $details;
+        $visitor['resolution'] = $this->getResolution();
     }
 
-    function getResolution()
+    protected function getResolution()
     {
         if (!array_key_exists('config_resolution', $this->details)) {
             return null;