+++ plugins/UserCountry/VisitorDetails.php
@@ -2,56 +2,62 @@
 /**
  * Piwik - free/libre analytics platform
  *
- * @link http://piwik.org
+ * @link    http://piwik.org
  * @license http://www.gnu.org/licenses/gpl-3.0.html GPL v3 or later
  *
  */
 namespace Piwik\Plugins\UserCountry;
 
-use Piwik\ArchiveProcessor;
 use Piwik\Common;
+use Piwik\Plugins\Live\VisitorDetailsAbstract;
 use Piwik\Plugins\UserCountry\LocationProvider\GeoIp;
-use Piwik\Plugins\UserCountry\LocationProvider;
 use Piwik\Tracker\Visit;
-use Piwik\Url;
 
 require_once PIWIK_INCLUDE_PATH . '/plugins/UserCountry/functions.php';
 
-class Visitor
+class VisitorDetails extends VisitorDetailsAbstract
 {
-    private $details = array();
-
-    public function __construct($details)
+    public function extendVisitorDetails(&$visitor)
     {
-        $this->details = $details;
+        $visitor['continent']     = $this->getContinent();
+        $visitor['continentCode'] = $this->getContinentCode();
+        $visitor['country']       = $this->getCountryName();
+        $visitor['countryCode']   = $this->getCountryCode();
+        $visitor['countryFlag']   = $this->getCountryFlag();
+        $visitor['region']        = $this->getRegionName();
+        $visitor['regionCode']    = $this->getRegionCode();
+        $visitor['city']          = $this->getCityName();
+        $visitor['location']      = $this->getPrettyLocation();
+        $visitor['latitude']      = $this->getLatitude();
+        $visitor['longitude']     = $this->getLongitude();
     }
 
-    public function getCountryCode()
+    protected function getCountryCode()
     {
         return $this->details['location_country'];
     }
 
-    public function getCountryName()
+    protected function getCountryName()
     {
         return countryTranslate($this->getCountryCode());
     }
 
-    public function getCountryFlag()
+    protected function getCountryFlag()
     {
         return getFlagFromCode($this->getCountryCode());
     }
 
-    public function getContinent()
+    protected function getContinent()
     {
         return continentTranslate($this->getContinentCode());
     }
 
-    public function getContinentCode()
+    protected function getContinentCode()
     {
         return Common::getContinent($this->details['location_country']);
     }
 
-    public function getCityName()
+    protected function getCityName()
     {
         if (!empty($this->details['location_city'])) {
             return $this->details['location_city'];
@@ -60,7 +66,7 @@ class Visitor
         return null;
     }
 
-    public function getRegionName()
+    protected function getRegionName()
     {
         $region = $this->getRegionCode();
         if ($region != '' && $region != Visit::UNKNOWN_CODE) {
@@ -71,12 +77,12 @@ class Visitor
         return null;
     }
 
-    public function getRegionCode()
+    protected function getRegionCode()
     {
         return $this->details['location_region'];
     }
 
-    public function getPrettyLocation()
+    protected function getPrettyLocation()
     {
         $parts = array();
 
@@ -95,7 +101,7 @@ class Visitor
         return implode(', ', $parts);
     }
 
-    public function getLatitude()
+    protected function getLatitude()
     {
         if (!empty($this->details['location_latitude'])) {
             return $this->details['location_latitude'];
@@ -104,7 +110,7 @@ class Visitor
         return null;
     }
 
-    public function getLongitude()
+    protected function getLongitude()
     {
         if (!empty($this->details['location_longitude'])) {
             return $this->details['location_longitude'];